Add detail page search nav component
